Move pythonhome_of to depsutils.jl and use it during build
