+++ src/pyinit.jl
@@ -90,59 +90,6 @@ _leak(T::Type, x::AbstractString) =
 _leak(::Type{Cwstring}, x::AbstractString) =
     _leak(Base.cconvert(Cwstring, x))
 
-function pythonhome_of(pyprogramname::AbstractString)
-    if Compat.Sys.iswindows()
-        script = """
-        import sys
-        if hasattr(sys, "base_exec_prefix"):
-            sys.stdout.write(sys.base_exec_prefix)
-        else:
-            sys.stdout.write(sys.exec_prefix)
-        """
-        # See where PYTHONHOME is mentioned in ../deps/build.jl
-    else
-        script = """
-        import sys
-        if hasattr(sys, "base_exec_prefix"):
-            sys.stdout.write(sys.base_prefix)
-            sys.stdout.write(":")
-            sys.stdout.write(sys.base_exec_prefix)
-        else:
-            sys.stdout.write(sys.prefix)
-            sys.stdout.write(":")
-            sys.stdout.write(sys.exec_prefix)
-        """
-        # https://docs.python.org/3/using/cmdline.html#envvar-PYTHONHOME
-    end
-    return read(python_cmd(`-c $script`; python = pyprogramname), String)
-end
-# To support `venv` standard library (as well as `virtualenv`), we
-# need to use `sys.base_prefix` and `sys.base_exec_prefix` here.
-# Otherwise, initializing Python in `__init__` below fails with
-# unrecoverable error:
-#
-#   Fatal Python error: initfsencoding: unable to load the file system codec
-#   ModuleNotFoundError: No module named 'encodings'
-#
-# This is because `venv` does not symlink standard libraries like
-# `virtualenv`.  For example, `lib/python3.X/encodings` does not
-# exist.  Rather, `venv` relies on the behavior of Python runtime:
-#
-#   If a file named "pyvenv.cfg" exists one directory above
-#   sys.executable, sys.prefix and sys.exec_prefix are set to that
-#   directory and it is also checked for site-packages
-#   --- https://docs.python.org/3/library/venv.html
-#
-# Thus, we need point `PYTHONHOME` to `sys.base_prefix` and
-# `sys.base_exec_prefix`.  If the virtual environment is created by
-# `virtualenv`, those `sys.base_*` paths point to the virtual
-# environment.  Thus, above code supports both use cases.
-#
-# See also:
-# * https://docs.python.org/3/library/venv.html
-# * https://docs.python.org/3/library/site.html
-# * https://docs.python.org/3/library/sys.html#sys.base_exec_prefix
-
 venv_python(::Nothing) = pyprogramname
 
 function venv_python(venv::AbstractString)
@@ -170,12 +117,7 @@ line arguments `args`.
 function python_cmd(args::Cmd = ``;
                     venv::Union{Nothing, AbstractString} = nothing,
                     python::AbstractString = venv_python(venv))
-    cmd = `$python $args`
-
-    # For Windows:
-    env = copy(ENV)
-    env["PYTHONIOENCODING"] = "UTF-8"
-    return setenv(cmd, env)
+    return pythonenv(`$python $args`)
 end
 
 function find_libpython(python::AbstractString)
